Improved the delete functionality

+++ lib/workouts_page.dart
@@ -5,6 +5,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:gym_day/checkin_form.dart';
+import 'package:flutter_progress_hud/flutter_progress_hud.dart';
+import 'package:modal_progress_hud/modal_progress_hud.dart';
 
 final _firestore = FirebaseFirestore.instance;
 final _auth = FirebaseAuth.instance;
@@ -17,23 +19,25 @@ class WorkoutsPage extends StatefulWidget {
 class _WorkoutsPageState extends State<WorkoutsPage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: StreamBuilder<QuerySnapshot>(
-        stream:
-            _firestore.collection('workouts').orderBy('dateTime').snapshots(),
-        builder: (context, snapshot) {
-          if (snapshot.hasData) {
-            final workouts = snapshot.data.docs;
-            for (var workout in workouts) {
-              print(workout.data()['workout']);
-            }
-            List<CustomWorkoutWidget> finalWorkouts = getWorkouts(workouts);
-            return ListView(
-              children: finalWorkouts,
-            );
-          } else
-            return Text('Error');
-        },
+    return ProgressHUD(
+      child: Scaffold(
+        body: StreamBuilder<QuerySnapshot>(
+          stream:
+              _firestore.collection('workouts').orderBy('dateTime').snapshots(),
+          builder: (context, snapshot) {
+            if (snapshot.hasData) {
+              final workouts = snapshot.data.docs;
+              for (var workout in workouts) {
+                print(workout.data()['workout']);
+              }
+              List<CustomWorkoutWidget> finalWorkouts = getWorkouts(workouts);
+              return ListView(
+                children: finalWorkouts,
+              );
+            } else
+              return Text('Error');
+          },
+        ),
       ),
     );
   }
@@ -44,20 +48,24 @@ class CustomWorkoutWidget extends StatefulWidget {
       {this.workoutName,
       this.workoutLocation,
       this.workoutDateTime,
-      this.workoutId});
+      this.workoutId,
+      });
 
   final String workoutName;
   final String workoutLocation;
   final Timestamp workoutDateTime;
   final String workoutId;
 
+
   @override
   _CustomWorkoutWidgetState createState() => _CustomWorkoutWidgetState();
 }
 
 class _CustomWorkoutWidgetState extends State<CustomWorkoutWidget> {
+
   @override
   Widget build(BuildContext context) {
+    final progress = ProgressHUD.of(context);
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 5.0),
       child: ListTile(
@@ -66,18 +74,49 @@ class _CustomWorkoutWidgetState extends State<CustomWorkoutWidget> {
           child: FaIcon(FontAwesomeIcons.times),
           minWidth: 10.0,
           onPressed: () async {
-            String deleteID;
-            await _firestore.collection('workouts').get().then((snapshot) {
-              snapshot.docs.forEach((doc) {
-                if (doc.data()['uid'] == widget.workoutId) {
-                  deleteID = doc.id;
-                }
-              });
-            });
+            showDialog(
+                context: context,
+                builder: (BuildContext context) {
+                  return AlertDialog(
+                    title: new Text('Alert'),
+                    content: new Text(
+                        'Are you sure you want to delete this workout?'),
+                    actions: <Widget>[
+                      // usually buttons at the bottom of the dialog
+                      new TextButton(
+                        child: new Text("Close"),
+                        onPressed: () {
+                          Navigator.of(context).pop();
+                        },
+                      ),
+                      new TextButton(
+                        child: new Text("Delete"),
+                        onPressed: () async {
+                          String deleteID;
+                          Navigator.pop(context);
+                          progress.showWithText('Deleting the workout');
+                          await _firestore
+                              .collection('workouts')
+                              .get()
+                              .then((snapshot) {
+                            snapshot.docs.forEach((doc) {
+                              if (doc.data()['uid'] == widget.workoutId) {
+                                deleteID = doc.id;
+                              }
+                            });
+                          });
 
-            await _firestore.collection('workouts').doc(deleteID).delete();
+                          await _firestore.collection('workouts').doc(deleteID).delete();
+                          // await Future.delayed(Duration(seconds: 10));
+                          progress.dismiss();
 
-            print('Deleted');
+
+                          print('Deleted');
+                        },
+                      ),
+                    ],
+                  );
+                });
           },
         ),
         title: Text(widget.workoutName),
